Step 5c: CallsTableQuery complete
